stable-diffusion-server => stable-diffusion-api
There were a couple of changes left to be made since the name change.
