+++ deployment/run_redis_app.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-poetry run uvicorn stable_diffusion_server.api.redis_app:app --host 0.0.0.0
+poetry run uvicorn stable_diffusion_api.api.redis_app:app --host 0.0.0.0
