+++ deployment/run_redis_worker.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-poetry run python -m stable_diffusion_server.engine.workers.redis_worker
+poetry run python -m stable_diffusion_api.engine.workers.redis_worker
