+++ stable_diffusion_api/api/in_memory_app.py
@@ -1,11 +1,11 @@
 import asyncio
 
-from stable_diffusion_server.api.base import AppConfig, create_app
-from stable_diffusion_server.engine.repos.blob_repo import InMemoryBlobRepo
-from stable_diffusion_server.engine.repos.key_value_repo import InMemoryKeyValueRepo
-from stable_diffusion_server.engine.repos.messaging_repo import InMemoryMessagingRepo
-from stable_diffusion_server.engine.repos.user_repo import InMemoryUserRepo
-from stable_diffusion_server.engine.workers.in_memory_worker import create_runner
+from stable_diffusion_api.api.base import AppConfig, create_app
+from stable_diffusion_api.engine.repos.blob_repo import InMemoryBlobRepo
+from stable_diffusion_api.engine.repos.key_value_repo import InMemoryKeyValueRepo
+from stable_diffusion_api.engine.repos.messaging_repo import InMemoryMessagingRepo
+from stable_diffusion_api.engine.repos.user_repo import InMemoryUserRepo
+from stable_diffusion_api.engine.workers.in_memory_worker import create_runner
 
 app_config = AppConfig(
     blob_repo_class=InMemoryBlobRepo,
