+++ stable_diffusion_api/api/tests/utils.py
