+++ stable_diffusion_api/engine/services/event_service.py
@@ -3,13 +3,13 @@ from typing import AsyncIterator, Optional
 
 import pydantic
 
-from stable_diffusion_server.engine.repos.blob_repo import BlobRepo
-from stable_diffusion_server.engine.repos.key_value_repo import KeyValueRepo
-from stable_diffusion_server.engine.repos.messaging_repo import MessagingRepo
-from stable_diffusion_server.engine.services.status_service import StatusService
-from stable_diffusion_server.engine.utils import _serialize_message, _deserialize_message
-from stable_diffusion_server.models.events import EventUnion
-from stable_diffusion_server.models.user import SessionId
+from stable_diffusion_api.engine.repos.blob_repo import BlobRepo
+from stable_diffusion_api.engine.repos.key_value_repo import KeyValueRepo
+from stable_diffusion_api.engine.repos.messaging_repo import MessagingRepo
+from stable_diffusion_api.engine.services.status_service import StatusService
+from stable_diffusion_api.engine.utils import _serialize_message, _deserialize_message
+from stable_diffusion_api.models.events import EventUnion
+from stable_diffusion_api.models.user import SessionId
 
 logger = logging.getLogger(__name__)
