+++ stable_diffusion_api/models/blob.py
