--- stable_diffusion_server/models/results.py
@@ -1,9 +1,0 @@
-import pydantic
-
-from stable_diffusion_server.models.blob import BlobUrl
-from stable_diffusion_server.models.params import ParamsUnion
-
-
-class GeneratedBlob(pydantic.BaseModel):
-    blob_url: BlobUrl
-    parameters_used: ParamsUnion
